refactor advisory workspace compare router
