+++ src/app/router_registry.py
@@ -55,6 +55,9 @@ from app.routers.advisory_workspace_actions import (
 from app.routers.advisory_workspace_assistant import (
     router as advisory_workspace_assistant_router,
 )
+from app.routers.advisory_workspace_compare import (
+    router as advisory_workspace_compare_router,
+)
 from app.routers.advisory_workspace_handoff import (
     router as advisory_workspace_handoff_router,
 )
@@ -403,6 +406,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(advisory_workspaces_router)
     app.include_router(advisory_workspace_actions_router)
     app.include_router(advisory_workspace_versions_router)
+    app.include_router(advisory_workspace_compare_router)
     app.include_router(advisory_workspace_version_lookups_router)
     app.include_router(advisory_workspace_assistant_router)
     app.include_router(advisory_workspace_handoff_router)
